ch(maintainability) set up function to handle repeated output formatting
